fix: make the project actually compile and pass its full CI pipeline locally
The 8-commit hardening series was verified with kotlinc stubs only; a real
AGP 8.13 + SDK 36 build surfaced real breakage. This commit fixes it all:

- ci.yml: testDebugUnitTest is ambiguous with flavors -> run both flavor
  test tasks explicitly (CI would have failed on the first push)
- KDoc: '/proc/*/environ' opens a NESTED block comment in Kotlin, leaving
  SovereignProxy.kt and RuntimeLaunchConfigBuilder.kt unparseable -> reword
- missing imports: AgentAutonomyMode (4 files), kotlinx CoroutineScope
- ClaudeRuntimeBridge: watchPermissionRequests is now suspend (delay)
- ConversationStore: setWriteAheadLoggingEnabled was removed in SDK 36 ->
  guarded enableWriteAheadLogging (public since 35; WAL default at 28)
- TaskResources: thermal listener renamed in SDK 36 -> reflective
  registration supporting both API names across runtimes
- RuntimeSetupService: restore the logFile() helper lost in window 0
- OutputFileTailer: wait for late-created output files instead of FNF
- build.gradle: unitTests.isReturnDefaultValues for AppLog in JVM tests
- i18n codemod damage in 3 UI files: hoist stringResource vals out of
  argument lists, restore missing argument commas, drop illegal commas
  inside if/else blocks, hoist resources out of a remember{} block,
  provide LocalContext in ProviderCredentialsStep
- tests: Socket port as Int (2 sites), ProviderKind.LLM_ROUTER name,
  missing parent mkdirs in WorkspaceCheckpointsTest

Verified locally on this machine (full pipeline, real SDK + NDK 26.1):
  unit tests: 468/468 green (234 x both flavors)
  assembleDebug: BUILD SUCCESSFUL, both APKs signed, native libs intact
  hygiene: size guard PASS, CycloneDX SBOM 7 components

+++ app/src/test/java/com/jarves/mh/runtime/SovereignProxyTest.kt
@@ -110,7 +110,7 @@ class SovereignProxyTest {
 
     /** Sends one request to the proxy and returns (status, body). */
     private fun request(proxy: SovereignProxy, path: String? = null, body: String = """{"m":"hi"}""", tokenOverride: String? = null): Pair<Int, String> {
-        Socket("127.0.0.1", proxy.url.substringAfterLast(":").substringBefore("/")).use { socket ->
+        Socket("127.0.0.1", proxy.url.substringAfterLast(":").substringBefore("/").toInt()).use { socket ->
             socket.soTimeout = 15_000
             val writer = BufferedWriter(OutputStreamWriter(socket.getOutputStream()))
             val target = path ?: "/t/${proxy.url.substringAfter("/t/")}/v1/messages"
@@ -211,7 +211,7 @@ class SovereignProxyTest {
             proxy.start()
             // Claim 40 MB without sending it: the proxy must answer 413 from
             // the header alone.
-            Socket("127.0.0.1", proxy.url.substringAfterLast(":").substringBefore("/")).use { socket ->
+            Socket("127.0.0.1", proxy.url.substringAfterLast(":").substringBefore("/").toInt()).use { socket ->
                 socket.soTimeout = 15_000
                 val writer = BufferedWriter(OutputStreamWriter(socket.getOutputStream()))
                 writer.write("POST /t/${proxy.url.substringAfter("/t/")}/v1/messages HTTP/1.1\r\n")
